requirements.txt
torch==2.8.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
datasets>=2.16.0
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.16.0
bitsandbytes>=0.41.0
safetensors>=0.4.0
sentencepiece>=0.1.99
tensorboard
deepspeed
torchvision
vllm==0.11.0
